requirements.txt
# --- Core app/runtime ---
streamlit==1.38.0
streamlit-drawable-canvas==0.9.3
pymupdf==1.24.9
Pillow==10.4.0

# --- Azure services ---
azure-ai-formrecognizer==3.3.2
azure-ai-textanalytics==5.3.0
# Optional but useful if you later switch to token creds/Managed Identity:
azure-identity==1.16.1

# --- LLM / Azure OpenAI client ---
openai==1.40.6

# --- NLP / matching utilities ---
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1

# --- Quality-of-life & data handling (used by the new code) ---
pandas==2.2.3
tqdm==4.66.5
python-dotenv==1.0.1